util: move all versions to a single file
