+++ tasks/util/versions.py
@@ -1,0 +1,18 @@
+# CoCo versions (note that the CoCo release pins the Kata Version)
+COCO_VERSION = "0.10.0"
+KATA_VERSION = "3.9.0"
+
+# Kubernetes versions
+K8S_VERSION = "1.28.2"
+CALICO_VERSION = "3.28.1"
+CNI_VERSION = "1.3.0"
+CRICTL_VERSION = "1.28.0"
+K9S_VERSION = "0.32.5"
+
+# Container image managament versions
+REGISTRY_VERSION = "2.8"
+COSIGN_VERSION = "2.2.0"
+SKOPEO_VERSION = "1.13.0"
+
+# Knative versions
+KNATIVE_VERSION = "1.15.0"
